fieldmanager: Move TestFieldManager into its own package
