+++ staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/lastappliedmanager_test.go
@@ -25,6 +25,7 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/fieldmanagertest"
 	"k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal"
 	"sigs.k8s.io/structured-merge-diff/v4/fieldpath"
 	"sigs.k8s.io/structured-merge-diff/v4/merge"
@@ -43,7 +44,7 @@ type testArgs struct {
 // created with the client-side apply last-applied annotation
 // will not give conflicts
 func TestApplyUsingLastAppliedAnnotation(t *testing.T) {
-	f := NewDefaultTestFieldManager(schema.FromAPIVersionAndKind("apps/v1", "Deployment"))
+	f := fieldmanagertest.NewDefaultTestFieldManager(schema.FromAPIVersionAndKind("apps/v1", "Deployment"))
 
 	tests := []testArgs{
 		{
@@ -563,7 +564,7 @@ spec:
 }
 
 func TestServiceApply(t *testing.T) {
-	f := NewDefaultTestFieldManager(schema.FromAPIVersionAndKind("v1", "Service"))
+	f := fieldmanagertest.NewDefaultTestFieldManager(schema.FromAPIVersionAndKind("v1", "Service"))
 
 	tests := []testArgs{
 		{
@@ -674,7 +675,7 @@ spec:
 }
 
 func TestReplicationControllerApply(t *testing.T) {
-	f := NewDefaultTestFieldManager(schema.FromAPIVersionAndKind("v1", "ReplicationController"))
+	f := fieldmanagertest.NewDefaultTestFieldManager(schema.FromAPIVersionAndKind("v1", "ReplicationController"))
 
 	tests := []testArgs{
 		{
@@ -737,7 +738,7 @@ spec:
 }
 
 func TestPodApply(t *testing.T) {
-	f := NewDefaultTestFieldManager(schema.FromAPIVersionAndKind("v1", "Pod"))
+	f := fieldmanagertest.NewDefaultTestFieldManager(schema.FromAPIVersionAndKind("v1", "Pod"))
 
 	tests := []testArgs{
 		{
@@ -914,7 +915,7 @@ spec:
 	testConflicts(t, f, tests)
 }
 
-func testConflicts(t *testing.T, f TestFieldManager, tests []testArgs) {
+func testConflicts(t *testing.T, f fieldmanagertest.TestFieldManager, tests []testArgs) {
 	for i, test := range tests {
 		t.Run(fmt.Sprintf("test %d", i), func(t *testing.T) {
 			f.Reset()
@@ -972,8 +973,8 @@ func testConflicts(t *testing.T, f TestFieldManager, tests []testArgs) {
 			}
 
 			// Eventually resource should contain applied changes
-			if !apiequality.Semantic.DeepDerivative(appliedObj, f.Get()) {
-				t.Errorf("expected equal resource: \n%#v, got: \n%#v", appliedObj, f.Get())
+			if !apiequality.Semantic.DeepDerivative(appliedObj, f.Live()) {
+				t.Errorf("expected equal resource: \n%#v, got: \n%#v", appliedObj, f.Live())
 			}
 		})
 	}
