+++ staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/lastappliedupdater_test.go
@@ -27,11 +27,12 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager"
+	"k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/fieldmanagertest"
 	"sigs.k8s.io/yaml"
 )
 
 func TestLastAppliedUpdater(t *testing.T) {
-	f := NewTestFieldManager(schema.FromAPIVersionAndKind("apps/v1", "Deployment"),
+	f := fieldmanagertest.NewTestFieldManager(schema.FromAPIVersionAndKind("apps/v1", "Deployment"),
 		"",
 		func(m fieldmanager.Manager) fieldmanager.Manager {
 			return fieldmanager.NewLastAppliedUpdater(m)
@@ -70,7 +71,7 @@ spec:
 		t.Errorf("error applying object: %v", err)
 	}
 
-	lastApplied, err := getLastApplied(f.liveObj)
+	lastApplied, err := getLastApplied(f.Live())
 	if err != nil {
 		t.Errorf("failed to get last applied: %v", err)
 	}
@@ -83,7 +84,7 @@ spec:
 		t.Errorf("error applying object: %v", err)
 	}
 
-	lastApplied, err = getLastApplied(f.liveObj)
+	lastApplied, err = getLastApplied(f.Live())
 	if err != nil {
 		t.Errorf("failed to get last applied: %v", err)
 	}
@@ -188,7 +189,7 @@ func TestLargeLastApplied(t *testing.T) {
 	}
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			f := NewTestFieldManager(schema.FromAPIVersionAndKind("v1", "ConfigMap"),
+			f := fieldmanagertest.NewTestFieldManager(schema.FromAPIVersionAndKind("v1", "ConfigMap"),
 				"",
 				func(m fieldmanager.Manager) fieldmanager.Manager {
 					return fieldmanager.NewLastAppliedUpdater(m)
@@ -198,7 +199,7 @@ func TestLargeLastApplied(t *testing.T) {
 				t.Errorf("Error applying object: %v", err)
 			}
 
-			lastApplied, err := getLastApplied(f.liveObj)
+			lastApplied, err := getLastApplied(f.Live())
 			if err != nil {
 				t.Errorf("Failed to access last applied annotation: %v", err)
 			}
@@ -211,12 +212,12 @@ func TestLargeLastApplied(t *testing.T) {
 			}
 
 			accessor := meta.NewAccessor()
-			annotations, err := accessor.Annotations(f.liveObj)
+			annotations, err := accessor.Annotations(f.Live())
 			if err != nil {
 				t.Errorf("Failed to access annotations: %v", err)
 			}
 			if annotations == nil {
-				t.Errorf("No annotations on obj: %v", f.liveObj)
+				t.Errorf("No annotations on obj: %v", f.Live())
 			}
 			lastApplied, ok := annotations[corev1.LastAppliedConfigAnnotation]
 			if ok || len(lastApplied) > 0 {
